- Move BONText to internal-only. - Add private testing pod for BONText. - Delete BONChainable protocol, and edit associated properties to just deal in BONChain. - Add some missing generic and nullable type declarations. - Change appendLink: to appendChain:. - Miscellaneous cleanup.
