+++ Example/BonMot/Cells/SpecialCharactersCell.m
@@ -44,19 +44,19 @@
 
     BONChain *baseTextChain = BONChain.new.textColor([UIColor darkGrayColor]);
     BONChain *baseImageChain = BONChain.new.baselineOffset(-10.0);
-    NSMutableArray *chunks = [NSMutableArray array];
+    BONGeneric(NSMutableArray, BONChain *)*chains = [NSMutableArray array];
 
     for (NSUInteger theIndex = 0; theIndex < imageNames.count; theIndex++) {
         UIImage *image = [UIImage imageNamed:imageNames[theIndex]];
         NSAssert(image, @"Image must not be nil");
-        BONChain *chunk = baseImageChain.image(image);
-        [chunk appendLink:baseTextChain.string(words[theIndex]) separator:BONSpecial.noBreakSpace];
+        BONChain *chain = baseImageChain.image(image);
+        [chain appendChain:baseTextChain.string(words[theIndex]) separator:BONSpecial.noBreakSpace];
 
-        [chunks addObject:chunk.text];
+        [chains addObject:chain];
     }
 
-    NSAttributedString *finalString = [BONText joinTexts:chunks
-                                           withSeparator:BONChain.new.string(BONSpecial.emSpace).text];
+    NSAttributedString *finalString = [BONChain joinChains:chains
+                                             withSeparator:BONChain.new.string(BONSpecial.emSpace)];
 
     self.label.attributedText = finalString;
