+++ Pod/BonMot.h
@@ -6,7 +6,6 @@
 //
 //
 
-#import "BONText.h"
 #import "BONChain.h"
 #import "BONSpecial.h"
 #import "UIImage+BonMotUtilities.h"
