--- Pod/Classes/BONText_Private.h
@@ -1,15 +1,0 @@
-//
-//  BONText.h
-//  BonMot
-//
-//  Created by Zev Eisenberg on 4/17/15.
-//
-//
-
-#import "BONText.h"
-
-@interface BONText ()
-
-@property (strong, nonatomic) NSNumber *internalIndentSpacer;
-
-@end
