+++ Pod/Classes/BONTypes.h
@@ -1,0 +1,21 @@
+//
+//  BONTypes.h
+//  BonMot
+//
+//  Created by Zev Eisenberg on 4/1/16.
+//
+//
+
+@import Foundation;
+
+typedef NS_ENUM(NSUInteger, BONFigureCase) {
+    BONFigureCaseDefault = 0,
+    BONFigureCaseLining,
+    BONFigureCaseOldstyle,
+};
+
+typedef NS_ENUM(NSUInteger, BONFigureSpacing) {
+    BONFigureSpacingDefault = 0,
+    BONFigureSpacingTabular,
+    BONFigureSpacingProportional,
+};
