+++ Pod/Classes/Private/BONChain_Private.h
@@ -1,0 +1,17 @@
+//
+//  BONChain_Private.h
+//  BonMot
+//
+//  Created by Zev Eisenberg on 4/1/16.
+//
+//
+
+#import "BONChain.h"
+#import "BONCompatibility.h"
+#import "BONText.h"
+
+@interface BONChain ()
+
+@property (BONNonnull, strong, nonatomic, readwrite) BONText *text;
+
+@end
