+++ Pod/UIKit/Classes/UILabel+BonMotUtilities.h
@@ -8,20 +8,20 @@
 
 @import UIKit;
 
-@protocol BONTextable;
+#import "BONCompatibility.h"
 
 @interface UILabel (BonMotUtilities)
 
 /**
- *  Assign a @c BONTextable object to apply to the label text. When a new value is assigned to @c bonString the chain attributes will be applied.
+ *  Assign a @c BONChain object to apply to the label text. When a new value is assigned to @c bonString the chain attributes will be applied.
  */
-@property (copy, nonatomic) id<BONTextable> bonTextable;
+@property (BONNullable, copy, nonatomic) BONChain *bonChain;
 
 /**
- *  Assigning a string via this method will apply the attributes of any assigned @c BONTextable and set the @c attributedText with the resulting @c NSAttributedString.
+ *  Assigning a string via this method will apply the attributes of any assigned @c bonChain and set the @c attributedText with the resulting @c NSAttributedString.
  *
  *  @param string The text to be displayed.
  */
-- (void)setBonString:(NSString *)string;
+- (void)setBonString:(BONNullable NSString *)string;
 
 @end
