+++ README.md
@@ -61,27 +61,9 @@ Think something is missing? Please [file an issue](https://github.com/Raizlabs/B
 
 ## Usage
 
-In any file where you want to use BonMot, simply `#import <BonMot/BonMot.h>` or `@import BonMot`.
+In any Swift file where you want to use BonMot, simply `import BonMot`. In Objective-C, that’s `#import <BonMot/BonMot.h>` or `@import BonMot`.
 
-The basic object in BonMot is `BONText`. You create a text object, set some properties to configure the font, and then ask for its `.attributedString` to get a string formatted according to your specification. Or ask for `.attributes` if you just need the attributes dictionary:
-
-```objc
-NSString *quote = @"I used to love correcting people’s grammar until\
- I realized what I loved more was having friends.\n\
-—Mara Wilson";
-
-BONText *text = [BONText new];
-text.font = [UIFont fontWithName:@"AmericanTypewriter" size:17.0];
-text.lineHeightMultiple = 1.8;
-text.string = quote;
-
-NSAttributedString *string = text.attributedString;
-NSDictionary *attributes = text.attributes;
-```
-
-## Chaining Syntax
-
-`BONChain` is a wrapper around `BONText` that allows you to chain properties together for a more concise expression of a style. You can create a chain with a normal `[[BONChain alloc] init]`, but it's easier to just use `[BONChain new]` or the even shorter and technically valid `BONChain.new`:
+The basic object in BonMot is `BONChain`, which allows you quickly construct attributed strings. You can create a chain with a normal `[[BONChain alloc] init]`, but it's easier to just use `[BONChain new]` or the even shorter and technically valid `BONChain.new`:
 
 ```objc
 NSString *quote = @"I used to love correcting people’s grammar until\
@@ -91,10 +73,11 @@ NSString *quote = @"I used to love correcting people’s grammar until\
 // line-wrapped for readability
 NSAttributedString *attributedString =
 BONChain.new // [BONChain new] and [[BONChain alloc] init] also work
-.fontNameAndSize(@"AmericanTypewriter", 17.0)
-.lineHeightMultiple(1.8)
-.string(quote)
-.attributedString;
+    .fontNameAndSize(@"AmericanTypewriter", 17.0)
+    .lineHeightMultiple(1.8)
+    .string(quote)
+    .attributedString; // You can also query .attributes
+                       // and get back a dictionary of attributes
 ```
 
 You can also create a local variable or property to save a partially-configured chain. All the chaining methods pass copies of the chain, so you don't have to worry about later changes clobbering earlier properties:
@@ -103,9 +86,9 @@ You can also create a local variable or property to save a partially-configured 
 // Base Chain
 BONChain *birdChain =
 BONChain.new
-.lineHeightMultiple(1.2)
-.font([UIFont systemFontOfSize:17.0])
-.string(@"bird");
+    .lineHeightMultiple(1.2)
+    .font([UIFont systemFontOfSize:17.0])
+    .string(@"bird");
 
 // Two chains with different colors
 // that inherit their parents’ properties
@@ -120,31 +103,31 @@ NSAttributedString *blueBirdString = blueBirds.attributedString;
 
 ## Concatenation
 
-You can concatenate an array of `BONText`s:
+You can concatenate an array of `BONChain`s:
 
 ```objc
-BONText *oneFish = BONChain.new.string(@"one fish").text;
-BONText *twoFish = BONChain.new.string(@"two fish").text;
-BONText *redFish = BONChain.new.string(@"red fish").textColor([UIColor redColor]).text;
-BONText *blueFish = BONChain.new.string(@"blue fish").textColor([UIColor blueColor]).text;
-BONText *separator = BONChain.new.string(@", ").text;
+BONChain *oneFish = BONChain.new.string(@"one fish");
+BONChain *twoFish = BONChain.new.string(@"two fish");
+BONChain *redFish = BONChain.new.string(@"red fish").textColor([UIColor redColor]);
+BONChain *blueFish = BONChain.new.string(@"blue fish").textColor([UIColor blueColor]);
+BONChain *separator = BONChain.new.string(@", ");
 
-NSAttributedString *string = [BONText joinTexts:@[ oneFish, twoFish, redFish, blueFish ] withSeparator:separator];
+NSAttributedString *string = [BONChain joinChains:@[ oneFish, twoFish, redFish, blueFish ] withSeparator:separator];
 ```
 
 Outputs:
 
 <img width=227 src="readme-images/fish-with-black-comma.png" />
 
-You can also append texts directly to each other:
+You can also append chains directly to each other:
 
 ```objc
 NSString *commaSpace = @", ";
 BONChain *chain = BONChain.new;
-[chain appendLink:BONChain.new.string(@"one fish")];
-[chain appendLink:BONChain.new.string(@"two fish") separator:commaSpace];
-[chain appendLink:BONChain.new.string(@"red fish").textColor([UIColor redColor]) separator:commaSpace];
-[chain appendLink:BONChain.new.string(@"blue fish").textColor([UIColor blueColor]) separator:commaSpace];
+[chain appendChain:BONChain.new.string(@"one fish")];
+[chain appendChain:BONChain.new.string(@"two fish") separator:commaSpace];
+[chain appendChain:BONChain.new.string(@"red fish").textColor([UIColor redColor]) separator:commaSpace];
+[chain appendChain:BONChain.new.string(@"blue fish").textColor([UIColor blueColor]) separator:commaSpace];
 
 NSAttributedString *string = chain.attributedString;
 ```
@@ -157,12 +140,12 @@ Outputs:
 
 ## Image Attachments
 
-BonMot uses `NSTextAttachment` to embed images in strings. Simply use the `.image` property of a chain or text:
+BonMot uses `NSTextAttachment` to embed images in strings. Simply use the `.image` property of a chain:
 
 ```objc
 BONChain *chain = BONChain.new;
-[chain appendLink:BONChain.new.image(someUIImage).baselineOffset(-4.0)];
-[chain appendLink:BONChain.new.string(@"label with icon") separator: @" "];
+[chain appendChain:BONChain.new.image(someUIImage).baselineOffset(-4.0)];
+[chain appendChain:BONChain.new.string(@"label with icon") separator: @" "];
 NSAttributedString *string = chain.attributedString;
 ```
 
@@ -175,8 +158,8 @@ If you need to wrap multiple lines of text after an image, use the `indentSpacer
 ```objc
 NSString *quote = @"This is some text that goes on and on and spans multiple lines, and it all ends up left-aligned";
 BONChain *chain = BONChain.new;
-[chain appendLink:BONChain.new.image(someUIImage).indentSpacer(10.0)];
-[chain appendLink:BONChain.new.string(quote)];
+[chain appendChain:BONChain.new.image(someUIImage).indentSpacer(10.0)];
+[chain appendChain:BONChain.new.string(quote)];
 NSAttributedString *attributedString = chain.attributedString;
 ```
 
@@ -196,7 +179,7 @@ BonMot provides several utilities that enhance its interoperability with UIKit.
 
 ### Text UI Elements
 
-BonMot provides a `bonTextable` property on `UILabel`, `UITextView`, and `UITextField` that allows assigning a `BONTextable` object to apply styling to any strings assigned via the `bonString` property.
+BonMot provides a `bonChain` property on `UILabel`, `UITextView`, and `UITextField` that allows assigning a `BONChain` object to apply styling to any strings assigned via the `bonString` property.
 
 **Note:** to use these utilities, add `pod 'BonMot/UIKit'` to your Podfile.
 
@@ -205,7 +188,7 @@ UILabel *label = [[UILabel alloc] init];
 
 BONChain *chain = BONChain.new.adobeTracking(300).fontNameAndSize(@"Avenir-Book", 18.0f);
 
-label.bonTextable = chain;
+label.bonChain = chain;
 label.bonString = @"Some initial text.";
 ```
 
@@ -213,7 +196,7 @@ Outputs:
 
 <img width=310 src="readme-images/initial-text.png" />
 
-Some time later, you can update the text of the label with a plain string, without losing the original styling from the `BONTextable`.
+Some time later, you can update the text of the label with a plain string, without losing the original styling from the `bonChain`.
 
 ```objc
 label.bonString = @"Some updated text.";
